refactor: operator controller pipeline_controller.go is very large, can be split into logical modules.
